1055 - expires_in is supposed to be Integer, so RefreshTokenResponse now accepts Integer instead of Long

+++ sdks/backend/java/component-api/src/main/java/com/bytechef/component/definition/Authorization.java
@@ -659,7 +659,7 @@ public interface Authorization {
      * @param accessToken
      * @param expiresIn
      */
-    record RefreshTokenResponse(String accessToken, String refreshToken, Long expiresIn) {
+    record RefreshTokenResponse(String accessToken, String refreshToken, Integer expiresIn) {
     }
 
     /**
