+++ server/libs/platform/platform-connection/platform-connection-api/src/main/java/com/bytechef/platform/connection/facade/ConnectionFacade.java
@@ -30,7 +30,7 @@ public interface ConnectionFacade {
 
     void delete(Long id);
 
-    Long executeConnectionRefresh(Long connectionId);
+    Integer executeConnectionRefresh(Long connectionId);
 
     ConnectionDTO getConnection(Long id);
